Add new method of core methods
