+++ Sources/CoreMethods/Public/UI/Fields/CardNumber/CardNumberTextField.swift
@@ -17,6 +17,11 @@ import UIKit
 ///
 /// Example usage:
 /// ```swift
+/// let style = TextFieldDefaultStyle()
+///    .borderColor(.systemGray)
+///    .borderWidth(2)
+///    .cornerRadius(8)
+///
 /// let field = CardNumberTextField(style: style)
 ///    .setMaxLength(19)
 ///    .setMask(pattern: "#### ##### ####")
@@ -24,9 +29,20 @@ import UIKit
 /// cardField.onBinChanged = { [weak self] bin in
 ///     // Handle BIN changes
 /// }
+///
+/// field.onLastFourDigitsFilled = { [weak self] lastFour in
+///     // Handle last four digits
+/// }
+///
+/// field.onFocusChanged = { [weak self] isFocus in
+///     // Handle focus changed
+/// }
+///
 /// cardField.onError = { [weak self] error in
 ///     // Handle validation errors
 /// }
+///
+/// field.setStyle(style) // Also you can change style this way
 /// ```
 public final class CardNumberTextField: PCITextField {
     /// Callback triggered when the BIN (Bank Identification Number) changes.
@@ -72,7 +88,12 @@ public final class CardNumberTextField: PCITextField {
     }
 
     // MARK: - Initialization
-
+    /// Initializer the textfield
+    ///
+    /// - Parameters:
+    ///   - style: The styling configuration for the text field
+    ///   - maxLength: Sets the maximum length of the card number
+    ///   - mask: Insert the mask pattern used for formatting the card number.
     public init(
         style: Style = TextFieldDefaultStyle(),
         maxLength: Int = 19,
