+++ Sources/CoreMethods/Public/UI/Fields/ExpirationDate/ExpirationDateTextfield.swift
@@ -17,6 +17,11 @@ import UIKit
 ///
 /// Example usage:
 /// ```swift
+/// let style = TextFieldDefaultStyle()
+///    .borderColor(.systemGray)
+///    .borderWidth(2)
+///    .cornerRadius(8)
+///    
 /// let field = ExpirationDateTextfield(style: style)
 ///    .setPlaceholder("Insert date")
 ///
@@ -29,6 +34,9 @@ import UIKit
 /// field.onInputFilled = { [weak self] error in
 ///     // Handle complete field
 /// }
+/// 
+/// field.setStyle(style) // Also you can change style this way
+///
 /// ```
 ///
 public final class ExpirationDateTextfield: PCITextField {
@@ -87,7 +95,11 @@ public final class ExpirationDateTextfield: PCITextField {
     }
 
     // MARK: - Initialization
-
+    
+    /// Initializer the textfield
+    ///
+    /// - Parameters:
+    ///   - style: The styling configuration for the text field
     public init(
         style: Style = TextFieldDefaultStyle()
     ) {
